+++ src/paging.rs
@@ -178,6 +178,29 @@ impl MemoryRegion {
     pub const fn is_empty(&self) -> bool {
         self.0.start.0 == self.0.end.0
     }
+
+    fn split(&self, level: usize) -> ChunkedIterator {
+        ChunkedIterator {
+            range: self,
+            granularity: granularity_at_level(level),
+            start: self.0.start.0,
+        }
+    }
+
+    /// Returns whether this region can be mapped at 'level' using block mappings only.
+    fn is_block(&self, level: usize) -> bool {
+        let gran = granularity_at_level(level);
+        (self.0.start.0 | self.0.end.0) & (gran - 1) == 0
+    }
+
+    /// Returns a new `MemoryRegion` based on this one but with the start aligned down and the end
+    /// aligned up to the given alignment.
+    fn align_out(&self, alignment: usize) -> Self {
+        Self(
+            VirtualAddress(align_down(self.0.start.0, alignment))
+                ..VirtualAddress(align_up(self.0.end.0, alignment)),
+        )
+    }
 }
 
 impl From<Range<VirtualAddress>> for MemoryRegion {
@@ -197,6 +220,24 @@ impl Debug for MemoryRegion {
         Display::fmt(self, f)
     }
 }
+
+/// A page table entry updater function; called repeatedly to update the state of a
+/// range of page table entries.
+///
+/// # Arguments
+///
+/// The updater function receives the following arguments:
+///
+/// - The full virtual address range mapped by the page table entry, which may be different than
+///   the original range passed to `modify_range`, due to alignment to block boundaries.
+/// - A page table entry whose state it may update.
+/// - The level of a translation table the entry belongs to.
+///
+/// # Return
+///
+/// - `Ok` to continue updating the remaining entries.
+/// - `Err` to signal an error during an update and stop updating the remaining entries.
+pub type PteUpdater = dyn Fn(&MemoryRegion, &mut Descriptor, usize) -> Result<(), ()>;
 
 /// A complete hierarchy of page tables including all levels.
 pub struct RootTable<T: Translation> {
@@ -234,7 +275,8 @@ impl<T: Translation> RootTable<T> {
     }
 
     /// Recursively maps a range into the pagetable hierarchy starting at the root level, mapping
-    /// the pages to the corresponding physical address range starting at `pa`.
+    /// the pages to the corresponding physical address range starting at `pa`. Block and page
+    /// entries will be written to, but will only be mapped if `flags` contains `Attributes::VALID`.
     ///
     /// Returns an error if the virtual address range is out of the range covered by the pagetable
     pub fn map_range(
@@ -243,28 +285,8 @@ impl<T: Translation> RootTable<T> {
         pa: PhysicalAddress,
         flags: Attributes,
     ) -> Result<(), MapError> {
-        if range.end() < range.start() {
-            return Err(MapError::RegionBackwards(range.clone()));
-        }
-        match self.va_range {
-            VaRange::Lower => {
-                if (range.start().0 as isize) < 0 {
-                    return Err(MapError::AddressRange(range.start()));
-                } else if range.end().0 > self.size() {
-                    return Err(MapError::AddressRange(range.end()));
-                }
-            }
-            VaRange::Upper => {
-                if range.start().0 as isize >= 0
-                    || (range.start().0 as isize).unsigned_abs() > self.size()
-                {
-                    return Err(MapError::AddressRange(range.start()));
-                }
-            }
-        }
-
+        self.verify_region(range)?;
         self.table.map_range(&self.translation, range, pa, flags);
-
         Ok(())
     }
 
@@ -283,6 +305,11 @@ impl<T: Translation> RootTable<T> {
         &self.translation
     }
 
+    pub fn modify_range(&mut self, range: &MemoryRegion, f: &PteUpdater) -> Result<(), MapError> {
+        self.verify_region(range)?;
+        self.table.modify_range(&self.translation, range, f)
+    }
+
     /// Returns the level of mapping used for the given virtual address:
     /// - `None` if it is unmapped
     /// - `Some(LEAF_LEVEL)` if it is mapped as a single page
@@ -290,6 +317,30 @@ impl<T: Translation> RootTable<T> {
     #[cfg(test)]
     pub(crate) fn mapping_level(&self, va: VirtualAddress) -> Option<usize> {
         self.table.mapping_level(&self.translation, va)
+    }
+
+    /// Checks whether the region is within range of the page table.
+    fn verify_region(&self, region: &MemoryRegion) -> Result<(), MapError> {
+        if region.end() < region.start() {
+            return Err(MapError::RegionBackwards(region.clone()));
+        }
+        match self.va_range {
+            VaRange::Lower => {
+                if (region.start().0 as isize) < 0 {
+                    return Err(MapError::AddressRange(region.start()));
+                } else if region.end().0 > self.size() {
+                    return Err(MapError::AddressRange(region.end()));
+                }
+            }
+            VaRange::Upper => {
+                if region.start().0 as isize >= 0
+                    || (region.start().0 as isize).unsigned_abs() > self.size()
+                {
+                    return Err(MapError::AddressRange(region.start()));
+                }
+            }
+        }
+        Ok(())
     }
 }
 
@@ -336,24 +387,9 @@ impl Iterator for ChunkedIterator<'_> {
     }
 }
 
-impl MemoryRegion {
-    fn split(&self, level: usize) -> ChunkedIterator {
-        ChunkedIterator {
-            range: self,
-            granularity: granularity_at_level(level),
-            start: self.0.start.0,
-        }
-    }
-
-    /// Returns whether this region can be mapped at 'level' using block mappings only.
-    fn is_block(&self, level: usize) -> bool {
-        let gran = granularity_at_level(level);
-        (self.0.start.0 | self.0.end.0) & (gran - 1) == 0
-    }
-}
-
 bitflags! {
     /// Attribute bits for a mapping in a page table.
+    #[derive(Copy, Clone, Debug, Eq, Hash, Ord, PartialEq, PartialOrd)]
     pub struct Attributes: usize {
         const VALID         = 1 << 0;
         const TABLE_OR_PAGE = 1 << 1;
@@ -367,7 +403,14 @@ bitflags! {
         const READ_ONLY     = 1 << 7;
         const ACCESSED      = 1 << 10;
         const NON_GLOBAL    = 1 << 11;
+        const DBM           = 1 << 51;
         const EXECUTE_NEVER = 3 << 53;
+
+        /// Software flags in block and page descriptor entries.
+        const SWFLAG_0 = 1 << 55;
+        const SWFLAG_1 = 1 << 56;
+        const SWFLAG_2 = 1 << 57;
+        const SWFLAG_3 = 1 << 58;
     }
 }
 
@@ -427,7 +470,8 @@ impl<T: Translation> PageTableWithLevel<T> {
     }
 
     /// Maps the the given virtual address range in this pagetable to the corresponding physical
-    /// address range starting at the given `pa`, recursing into any subtables as necessary.
+    /// address range starting at the given `pa`, recursing into any subtables as necessary. To map
+    /// block and page entries, `Attributes::VALID` must be set in `flags`.
     ///
     /// Assumes that the entire range is within the range covered by this pagetable.
     ///
@@ -464,19 +508,22 @@ impl<T: Translation> PageTableWithLevel<T> {
                 } else {
                     let old = *entry;
                     let (mut subtable, subtable_pa) = Self::new(translation, level + 1);
-                    if let (Some(old_flags), Some(old_pa)) = (old.flags(), old.output_address()) {
-                        // Old was a valid block entry, so we need to split it.
-                        // Recreate the entire block in the newly added table.
-                        let a = align_down(chunk.0.start.0, granularity);
-                        let b = align_up(chunk.0.end.0, granularity);
-                        subtable.map_range(
-                            translation,
-                            &MemoryRegion::new(a, b),
-                            old_pa,
-                            old_flags,
-                        );
+                    if let Some(old_flags) = old.flags() {
+                        if !old_flags.contains(Attributes::TABLE_OR_PAGE) {
+                            let old_pa = old.output_address();
+                            // `old` was a block entry, so we need to split it.
+                            // Recreate the entire block in the newly added table.
+                            let a = align_down(chunk.0.start.0, granularity);
+                            let b = align_up(chunk.0.end.0, granularity);
+                            subtable.map_range(
+                                translation,
+                                &MemoryRegion::new(a, b),
+                                old_pa,
+                                old_flags,
+                            );
+                        }
                     }
-                    entry.set(subtable_pa, Attributes::TABLE_OR_PAGE);
+                    entry.set(subtable_pa, Attributes::TABLE_OR_PAGE | Attributes::VALID);
                     subtable
                 };
                 subtable.map_range(translation, &chunk, pa, flags);
@@ -491,6 +538,7 @@ impl<T: Translation> PageTableWithLevel<T> {
         translation: &T,
         indentation: usize,
     ) -> Result<(), fmt::Error> {
+        const WIDTH: usize = 3;
         // Safe because we know that the pointer is aligned, initialised and dereferencable, and the
         // PageTable won't be mutated while we are using it.
         let table = unsafe { self.table.as_ref() };
@@ -503,12 +551,16 @@ impl<T: Translation> PageTableWithLevel<T> {
                     i += 1;
                 }
                 if i - 1 == first_zero {
-                    writeln!(f, "{:indentation$}{}: 0", "", first_zero)?;
+                    writeln!(f, "{:indentation$}{: <WIDTH$}: 0", "", first_zero)?;
                 } else {
-                    writeln!(f, "{:indentation$}{}-{}: 0", "", first_zero, i - 1)?;
+                    writeln!(f, "{:indentation$}{: <WIDTH$}-{}: 0", "", first_zero, i - 1)?;
                 }
             } else {
-                writeln!(f, "{:indentation$}{}: {:?}", "", i, table.entries[i])?;
+                writeln!(
+                    f,
+                    "{:indentation$}{: <WIDTH$}: {:?}",
+                    "", i, table.entries[i],
+                )?;
                 if let Some(subtable) = table.entries[i].subtable(translation, self.level) {
                     subtable.fmt_indented(f, translation, indentation + 2)?;
                 }
@@ -537,6 +589,28 @@ impl<T: Translation> PageTableWithLevel<T> {
             // Actually free the memory used by the `PageTable`.
             translation.deallocate_table(self.table);
         }
+    }
+
+    /// Modifies a range of page table entries by applying a function to each page table entry.
+    /// If the range is not aligned to block boundaries, it will be expanded.
+    fn modify_range(
+        &mut self,
+        translation: &T,
+        range: &MemoryRegion,
+        f: &PteUpdater,
+    ) -> Result<(), MapError> {
+        let level = self.level;
+        for chunk in range.split(level) {
+            // VA range passed to the updater is aligned to block boundaries, as that region will
+            // be affected by changes to the entry.
+            let affected_range = chunk.align_out(granularity_at_level(level));
+            let entry = self.get_entry_mut(chunk.0.start);
+            f(&affected_range, entry, level).map_err(|_| MapError::PteUpdateFault(*entry))?;
+            if let Some(mut subtable) = entry.subtable(translation, level) {
+                subtable.modify_range(translation, &chunk, f)?;
+            }
+        }
+        Ok(())
     }
 
     /// Returns the level of mapping used for the given virtual address:
@@ -582,55 +656,55 @@ impl PageTable {
 ///   - A page mapping, if it is in the lowest level page table.
 ///   - A block mapping, if it is not in the lowest level page table.
 ///   - A pointer to a lower level pagetable, if it is not in the lowest level page table.
-#[derive(Clone, Copy)]
+#[derive(Clone, Copy, PartialEq, Eq)]
 #[repr(C)]
-struct Descriptor(usize);
+pub struct Descriptor(usize);
 
 impl Descriptor {
-    fn output_address(&self) -> Option<PhysicalAddress> {
-        if self.is_valid() {
-            Some(PhysicalAddress(
-                self.0 & (!(PAGE_SIZE - 1) & !(0xffff << 48)),
-            ))
-        } else {
-            None
-        }
+    const PHYSICAL_ADDRESS_BITMASK: usize = !(PAGE_SIZE - 1) & !(0xffff << 48);
+
+    fn output_address(self) -> PhysicalAddress {
+        PhysicalAddress(self.0 & Self::PHYSICAL_ADDRESS_BITMASK)
     }
 
-    fn flags(self) -> Option<Attributes> {
-        if self.is_valid() {
-            Attributes::from_bits(self.0 & ((PAGE_SIZE - 1) | (0xffff << 48)))
-        } else {
-            None
-        }
+    /// Returns the flags of this page table entry, or `None` if its state does not
+    /// contain a valid set of flags.
+    pub fn flags(self) -> Option<Attributes> {
+        Attributes::from_bits(self.0 & !Self::PHYSICAL_ADDRESS_BITMASK)
     }
 
-    fn is_valid(self) -> bool {
+    /// Modifies the page table entry by setting or clearing its flags.
+    pub fn modify_flags(&mut self, set: Attributes, clear: Attributes) {
+        self.0 = (self.0 | set.bits()) & !clear.bits();
+    }
+
+    /// Returns `true` if [`Attributes::VALID`] is set on this entry, e.g. if the entry is mapped.
+    pub fn is_valid(self) -> bool {
         (self.0 & Attributes::VALID.bits()) != 0
     }
 
-    fn is_table_or_page(self) -> bool {
+    /// Returns `true` if this is a valid entry pointing to a next level translation table or a page.
+    pub fn is_table_or_page(self) -> bool {
         if let Some(flags) = self.flags() {
-            flags.contains(Attributes::TABLE_OR_PAGE)
+            flags.contains(Attributes::TABLE_OR_PAGE | Attributes::VALID)
         } else {
             false
         }
     }
 
     fn set(&mut self, pa: PhysicalAddress, flags: Attributes) {
-        self.0 = pa.0 | (flags | Attributes::VALID).bits();
+        self.0 = (pa.0 & Self::PHYSICAL_ADDRESS_BITMASK) | flags.bits();
     }
 
     fn subtable<T: Translation>(
-        &self,
+        self,
         translation: &T,
         level: usize,
     ) -> Option<PageTableWithLevel<T>> {
         if level < LEAF_LEVEL && self.is_table_or_page() {
-            if let Some(output_address) = self.output_address() {
-                let table = translation.physical_to_virtual(output_address);
-                return Some(PageTableWithLevel::from_pointer(table, level + 1));
-            }
+            let output_address = self.output_address();
+            let table = translation.physical_to_virtual(output_address);
+            return Some(PageTableWithLevel::from_pointer(table, level + 1));
         }
         None
     }
@@ -639,8 +713,10 @@ impl Descriptor {
 impl Debug for Descriptor {
     fn fmt(&self, f: &mut Formatter) -> Result<(), fmt::Error> {
         write!(f, "{:#016x}", self.0)?;
-        if let (Some(flags), Some(address)) = (self.flags(), self.output_address()) {
-            write!(f, " ({}, {:?})", address, flags)?;
+        if self.is_valid() {
+            if let Some(flags) = self.flags() {
+                write!(f, " ({}, {:?})", self.output_address(), flags)?;
+            }
         }
         Ok(())
     }
@@ -691,7 +767,7 @@ pub(crate) const fn is_aligned(value: usize, alignment: usize) -> bool {
 mod tests {
     use super::*;
     #[cfg(feature = "alloc")]
-    use alloc::{format, string::ToString};
+    use alloc::{format, string::ToString, vec, vec::Vec};
 
     #[cfg(feature = "alloc")]
     #[test]
@@ -751,5 +827,62 @@ mod tests {
     #[test]
     fn add_physical_address() {
         assert_eq!(PhysicalAddress(0x1234) + 0x42, PhysicalAddress(0x1276));
+    }
+
+    #[test]
+    fn invalid_descriptor() {
+        let desc = Descriptor(0usize);
+        assert!(!desc.is_valid());
+        assert!(!desc.flags().unwrap().contains(Attributes::VALID));
+    }
+
+    #[test]
+    fn set_descriptor() {
+        const PHYSICAL_ADDRESS: usize = 0x12340000;
+        let mut desc = Descriptor(0usize);
+        assert!(!desc.is_valid());
+        desc.set(
+            PhysicalAddress(PHYSICAL_ADDRESS),
+            Attributes::TABLE_OR_PAGE | Attributes::USER | Attributes::SWFLAG_1 | Attributes::VALID,
+        );
+        assert!(desc.is_valid());
+        assert_eq!(
+            desc.flags().unwrap(),
+            Attributes::TABLE_OR_PAGE | Attributes::USER | Attributes::SWFLAG_1 | Attributes::VALID
+        );
+        assert_eq!(desc.output_address(), PhysicalAddress(PHYSICAL_ADDRESS));
+    }
+
+    #[test]
+    fn modify_descriptor_flags() {
+        let mut desc = Descriptor(0usize);
+        assert!(!desc.is_valid());
+        desc.set(
+            PhysicalAddress(0x12340000),
+            Attributes::TABLE_OR_PAGE | Attributes::USER | Attributes::SWFLAG_1,
+        );
+        desc.modify_flags(
+            Attributes::DBM | Attributes::SWFLAG_3,
+            Attributes::VALID | Attributes::SWFLAG_1,
+        );
+        assert!(!desc.is_valid());
+        assert_eq!(
+            desc.flags().unwrap(),
+            Attributes::TABLE_OR_PAGE | Attributes::USER | Attributes::SWFLAG_3 | Attributes::DBM
+        );
+    }
+
+    #[cfg(feature = "alloc")]
+    #[test]
+    fn unaligned_chunks() {
+        let region = MemoryRegion::new(0x0000_2000, 0x0020_5000);
+        let chunks = region.split(LEAF_LEVEL - 1).collect::<Vec<_>>();
+        assert_eq!(
+            chunks,
+            vec![
+                MemoryRegion::new(0x0000_2000, 0x0020_0000),
+                MemoryRegion::new(0x0020_0000, 0x0020_5000),
+            ]
+        );
     }
 }